Changed x/y to posX/posY in agents

+++ src/main/java/interfaces/TrafficListener.java
@@ -3,11 +3,11 @@ package interfaces;
 import utils.LightColor;
 
 public interface TrafficListener {
-    void spawnTrafficLight(boolean isGreen, int trafficLightId, double initialX, double initialY);
+    void spawnTrafficLight(boolean isGreen, int trafficLightId, int initialX, int initialY);
 
     void updateTrafficLight(int trafficLightId, LightColor color);
 
-    void spawnCar(int carId, double initialX, double initialY);
+    void spawnCar(int carId, int initialX, int initialY);
 
     void moveCar(int carId, double newX, double newY, double angle);
